Mechanical change: move kit-config code into separate bundle and serve files. No code changes.
