+++ packages/cli/src/bundle/kit-config.ts
@@ -1,12 +1,8 @@
-import type {
-  BundleDefinitionWithRequiredParameters,
-  ServerWithRequiredParameters,
-} from "@rnx-kit/config";
+import type { BundleDefinitionWithRequiredParameters } from "@rnx-kit/config";
 import {
   getBundleDefinition,
   getBundlePlatformDefinition,
   getKitConfig,
-  getServerConfig,
 } from "@rnx-kit/config";
 import { warn } from "@rnx-kit/console";
 import type { BundleArgs } from "@rnx-kit/metro-service";
@@ -99,33 +95,5 @@ export function getKitBundlePlatformDefinition(
         "experimental_treeShake",
       ]
     ),
-  };
-}
-
-export type ServerConfigOverrides = {
-  projectRoot?: string;
-  assetPlugins?: string[];
-  sourceExts?: string[];
-};
-
-/**
- * Get the kit's server configuration. Apply any overrides.
- *
- * @param overrides Overrides to apply to the output server configuration. These take precedence over the values in the kit config.
- * @returns Overridden server configuration
- */
-export function getKitServerConfig(
-  overrides: ServerConfigOverrides
-): ServerWithRequiredParameters {
-  const kitConfig = getKitConfig();
-  if (!kitConfig) {
-    throw new Error(
-      "No kit configuration found for this react-native experience"
-    );
-  }
-
-  return {
-    ...getServerConfig(kitConfig),
-    ...pickValues(overrides, ["projectRoot", "assetPlugins", "sourceExts"]),
   };
 }
